submissions/extras/requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cycler==0.10.0
idna==2.10
joblib==1.0.0
kiwisolver==1.3.1
matplotlib==3.3.3
nltk==3.5
numpy==1.19.3
oauthlib==3.1.0
pandas==1.2.0
Pillow==8.0.1
pymongo==3.11.2
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.5
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
textblob==0.15.3
tqdm==4.55.0
tweepy==3.10.0
urllib3==1.26.2
wordcloud==1.8.1
